requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langchain-qdrant
qdrant-client
langgraph
google-generativeai==0.3.1
pypdf2==3.0.1
typing-extensions>=4.9.0  # Updated from 4.8.0 to 4.9.0 or newer
cors==1.0.1